Finally setup a way to properly have test without messing with `module-info.java` Now the files are structured in a modular way. Java sources are located in `src/main/module/package` Test sources are located in `src/test/module/package` Note that we don't need different module and/or package for java sources & their corresponding test
