+++ src/main/cx.ksim.dicelock/module-info.java
@@ -1,0 +1,4 @@
+
+module cx.ksim.dicelock {
+	
+}
