--- src/module-info.java
@@ -1,9 +1,0 @@
-/**
- * 
- */
-/**
- * @author Admin
- *
- */
-module DiceLock {
-}
