--- test/cx/ksim/dicelock/DiceLock_test.java
@@ -1,14 +1,0 @@
-package cx.ksim.dicelock;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import org.junit.jupiter.api.Test;
-
-class DiceLock_test {
-
-	@Test
-	void testMain() {
-		fail( "Not yet implemented" );
-	}
-
-}
